Eleven
The program shows "lose" when I actually win the game, I check the condition which the board should be all null and deck should have length of 0. Both of those are true at the end of the game, which means that GameIsWon is true, but it still shows lose. I don't know how to deal with it further.